% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.translate_operations.R
\name{delete_terminology}
\alias{delete_terminology}
\title{A synchronous action that deletes a custom terminology}
\usage{
delete_terminology(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the custom terminology being deleted.}
}
\description{
A synchronous action that deletes a custom terminology.
}
\examples{

}
