% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.translate_operations.R
\name{get_terminology}
\alias{get_terminology}
\title{Retrieves a custom terminology}
\usage{
get_terminology(
  Name = "string",
  TerminologyDataFormat = "CSV"|"TMX"
)
}
\arguments{
\item{Name}{[required] The name of the custom terminology being retrieved.}

\item{TerminologyDataFormat}{[required] The data format of the custom terminology being retrieved, either CSV or TMX.}
}
\description{
Retrieves a custom terminology.
}
\examples{

}
