% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.translate_operations.R
\name{list_terminologies}
\alias{list_terminologies}
\title{Provides a list of custom terminologies associated with your account}
\usage{
list_terminologies(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{If the result of the request to ListTerminologies was truncated, include the NextToken to fetch the next group of custom terminologies.}

\item{MaxResults}{The maximum number of custom terminologies returned per list request.}
}
\description{
Provides a list of custom terminologies associated with your account.
}
\section{Accepted Parameters}{
\preformatted{list_terminologies(
  NextToken = "string",
  MaxResults = 123
)
}
}

