% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.translate_package.R
\docType{package}
\name{paws.translate-package}
\alias{paws.translate}
\alias{paws.translate-package}
\title{paws.translate: Amazon Translate}
\description{
Provides translation between English and one of six
languages, or between one of the six languages and English.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
