% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.translate_package.R
\docType{package}
\name{paws.translate-package}
\alias{paws.translate}
\alias{paws.translate-package}
\title{paws.translate: Amazon Translate}
\description{
Provides translation between one source language and another
of the same set of languages.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
