% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.translate_operations.R
\name{translate_text}
\alias{translate_text}
\title{Translates input text from the source language to the target language}
\usage{
translate_text(Text, SourceLanguageCode, TargetLanguageCode)
}
\arguments{
\item{Text}{[required] The text to translate.}

\item{SourceLanguageCode}{[required] One of the supported language codes for the source text. If the \code{TargetLanguageCode} is not "en", the \code{SourceLanguageCode} must be "en".

To have Amazon Translate determine the source language of your text, you can specify \code{auto} in the \code{SourceLanguageCode} field. If you specify \code{auto}, Amazon Translate will call Amazon Comprehend to determine the source language.}

\item{TargetLanguageCode}{[required] One of the supported language codes for the target text. If the \code{SourceLanguageCode} is not "en", the \code{TargetLanguageCode} must be "en".}
}
\description{
Translates input text from the source language to the target language. You can translate between English (en) and one of the following languages, or between one of the following languages and English.
}
\details{
\itemize{
\item Arabic (ar)
\item Chinese (Simplified) (zh)
\item French (fr)
\item German (de)
\item Portuguese (pt)
\item Spanish (es)
}

To have Amazon Translate determine the source language of your text, you can specify \code{auto} in the \code{SourceLanguageCode} field. If you specify \code{auto}, Amazon Translate will call Amazon Comprehend to determine the source language.
}
\section{Accepted Parameters}{
\preformatted{translate_text(
  Text = "string",
  SourceLanguageCode = "string",
  TargetLanguageCode = "string"
)
}
}

